move SynthesisEngine mora utility under synthesis_engine
