+++ voicevox_engine/synthesis_engine/SynthesisEngine.py
@@ -6,9 +6,9 @@ from scipy.signal import resample
 from ..acoustic_feature_extractor import OjtPhoneme, SamplingData
 from ..full_context_label import extract_full_context_label
 from ..model import AccentPhrase, AudioQuery, Mora
-from ..mora.mora_to_text import mora_to_text
-from ..mora.pre_process import pre_process
-from ..mora.split_mora import split_mora, unvoiced_mora_phoneme_list
+from .mora.mora_to_text import mora_to_text
+from .mora.pre_process import pre_process
+from .mora.split_mora import split_mora, unvoiced_mora_phoneme_list
 
 
 class SynthesisEngine:
