+++ voicevox_engine/synthesis_engine/mora/split_mora.py
@@ -1,6 +1,6 @@
 from typing import List, Optional
 
-from ..acoustic_feature_extractor import OjtPhoneme
+from ...acoustic_feature_extractor import OjtPhoneme
 
 unvoiced_mora_phoneme_list = ["A", "I", "U", "E", "O", "cl", "pau"]
 mora_phoneme_list = ["a", "i", "u", "e", "o", "N"] + unvoiced_mora_phoneme_list
